ldbl-128ibm-compat: Test double values
A single format string can take double and long double parameters at the
same time.  Internally, these parameters are routed to the same
function, which correctly reads them and calls the underlying functions
responsible for the actual conversion to string.  This patch adds a new
case to test this scenario.

Tested for powerpc64le.

	* sysdeps/ieee754/ldbl-128ibm-compat/test-printf-ldbl-compat.c:
	(do_test_call_rarg): Add parameter and use it in the calls
	to *printf functions under test.
	(do_test): Add variable and use it in the calls to
	do_test_call_rarg and do_test_call_varg.
	* sysdeps/ieee754/ldbl-128ibm-compat/test-wprintf-ldbl-compat.c:
	Likewise.
	* sysdeps/ieee754/ldbl-128ibm-compat/test-printf-ldbl-compat.sh:
	Modify expected result.
	* sysdeps/ieee754/ldbl-128ibm-compat/test-wprintf-ldbl-compat.sh:
	Likewise.