ldbl-128ibm-compat: Add printf_size
Since the addition of the _Float128 API, strfromf128 and printf_size use
__printf_fp to print _Float128 values.  This is achieved by setting the
'is_binary128' member of the 'printf_info' structure to 1.  Now that the
format of long double on powerpc64le is getting a third option, this
mechanism is reused for long double values that have binary128 format
(i.e.: when -mabi=ieeelongdouble).

This patch adds __printf_sizeieee128 as an exported symbol, but doesn't
provide redirections from printf_size, yet.  All redirections will be
installed in a future commit, once all other functions that print or
read long double values with binary128 format are ready.

Tested for powerpc64le.

	* sysdeps/ieee754/ldbl-128ibm-compat/Makefile:
	[subdir == stdio-common] (routines): Add ieee128-printf_size.
	[subdir == stdio-common] (tests-internal): Add
	test-printf-size-ieee128, and test-printf-size-ibm128.
	[subdir == stdio-common] (CFLAGS-test-printf-size-ieee128.c)
	(CFLAGS-test-printf-size-ibm128.c): New variables.
	[subdir == stdio-common] (tests-special): Add
	$(objpfx)test-printf-size-ieee128.out and
	$(objpfx)test-printf-size-ibm128.out.
	[subdir == stdio-common] ($(objpfx)test-printf-size-ieee128.out)
	($(objpfx)test-printf-size-ibm128.out): New build and run rules.
	* sysdeps/ieee754/ldbl-128ibm-compat/Versions (libc): Add
	__printf_sizeieee128.
	* sysdeps/ieee754/ldbl-128ibm-compat/ieee128-printf_size.c:
	New file.
	* sysdeps/ieee754/ldbl-128ibm-compat/test-printf-size-ibm128.c:
	Likewise.
	* sysdeps/ieee754/ldbl-128ibm-compat/test-printf-size-ieee128.c:
	Likewise.