Fix behaviour of 'is_binary128' in __printf_fp
Should this have an associated bugzilla entry?

-- 8< --
Both 'is_long_double' and 'is_binary128' (members of the 'printf_info'
structure) are used in __printf_fp to select the type of the
floating-point argument it receives in its 'args' parameter.  When the
argument is of double type, 'is_long_double' should be set to zero.
Similarly, when the argument is of long double type, 'is_long_double'
should be set to one.  More recently (since the addition of the
_Float128 API), the floating-point argument can also be of _Float128
type, then 'is_binary128' should be set to one.

In current __printf_fp, when 'is_binary128' is one, the floating-point
argument is treated as if it was of _Float128 type, regardless of the
value of 'is_long_double'.  This is not a problem for strfromd, because
strfromd always sets 'is_binary128' to zero.  However, it could be a
problem for users of printf_size, who could be expecting that
'is_long_double == 0' meant that the 'args' parameter was of double
type, regardless of the value of 'is_binary128'.

This patch modifies this behaviour in __printf_fp and adjusts
strfromf128 so that it always sets 'is_long_double' to one.

Tested for powerpc64le and x86_64.

	* stdio-common/printf_fp.c (__printf_fp_l): Also check the value
	of 'is_long_double' before reading the argument into a variable
	of _Float128 type.
	* stdio-common/printf_size.c (__printf_size): Likewise.
	* stdlib/strfrom-skeleton.c [__HAVE_DISTINCT_FLOAT128] (STRFROM):
	If 'is_binary128' is one, also set 'is_long_double' to one.

	* stdio-common/Makefile (test-srcs): Add tst-isbinary128
	(tests-special) Add $(objpfx)tst-isbinary128.out.
	($(objpfx)tst-isbinary128.out): New build and run rule.
	* stdio-common/tst-isbinary128.c: New file.
	* stdio-common/tst-isbinary128.sh: Likewise.